+++ kaparoo/utils/__init__.py
@@ -1,7 +1,16 @@
 __all__ = (
+    "Aggregator",
+    "Fold",
+    "Last",
+    "Max",
+    "Mean",
+    "Min",
+    "Reduction",
     "SegmentRecord",
     "SegmentTimer",
+    "Sum",
     "Timer",
+    "UnweightedReduction",
     "factory_if_none",
     "replace_if_none",
     "unwrap_or_default",
@@ -10,6 +19,17 @@ __all__ = (
     "unwrap_or_factory",
 )
 
+from kaparoo.utils.aggregate import (
+    Aggregator,
+    Fold,
+    Last,
+    Max,
+    Mean,
+    Min,
+    Reduction,
+    Sum,
+    UnweightedReduction,
+)
 from kaparoo.utils.optional import (
     factory_if_none,
     replace_if_none,